+++ R/find_random.R
@@ -40,35 +40,22 @@
 #' }
 #' @export
 find_random <- function(x, split_nested = FALSE, flatten = FALSE) {
+  f <- find_formula(x)
 
-  # Deal with GAMM objects
-  is_gam <- TRUE  # Flag to later remove dummy random
-  if (inherits(x, "gamm")) {
-    l <- list("random" = names(x$lme$groups))
-  } else if (inherits(x, "list") && "mer" %in% names(x)) {
-    x <- x$mer
-    l <- .find_random(x, split_nested)
-  # Deal with all other models
+  if (is_multivariate(x)) {
+    rn <- names(find_response(x))
+    l <- lapply(rn, function(i) .find_random_effects(x, f[[i]], split_nested))
+    names(l) <- rn
+    l <- .compact_list(l)
   } else {
-    is_gam <- FALSE
-    l <- .find_random(x, split_nested)
+    l <- .find_random_effects(x, f, split_nested)
   }
 
-  # Sanitize empty return
+
   if (.is_empty_object(l)) {
     return(NULL)
   }
 
-  # Clean output from GAMMs (https://github.com/easystats/insight/issues/42)
-  if (is_gam) {
-    if (any(c("Xr.0", "g.0") %in% l$random)) {
-      l$random <- l$random[!l$random %in% c("Xr.0", "g.0")]
-    } else{
-      l$random <- l$random[!l$random %in% c("Xr", "g")]
-    }
-  }
-
-  # Prepare output
   if (flatten) {
     unique(unlist(l))
   } else {